Add ability to retrieve a list of test nodes
Summary:
This diff adds the ability to traverse the list view and build `TestNodes` via their `Components`. I have copied code we already have in our current testing API and adapted to allow to recursively search for hierarchy in sections/collections.

This is crucial so that we can actually perform full tests using this new API and validate that other aspects of the framework work (matchers, interactions, etc).

Reviewed By: pentiumao

Differential Revision: D41681999

fbshipit-source-id: 47b36fcfad355340175ac973134e3845cdcf455c