Clones the Drawable state to avoid crashes due to race conditions
Summary:
Clones the DrawableState before getting the padding during layout to avoid crashes due to race conditions since the same instance of the Drawable could be currently mounted (from a previous layout pass).

This is only a problem when render is split into resolve and layout, and litho reuses the previous ResolvedResult to calculate a new layout. In the normal case the Drawable would always be resolved again.

Reviewed By: astreet

Differential Revision: D41720891

fbshipit-source-id: c69e80c558c3f44eb0997092d306a43e93a27f65