Purge MeasuredResultCache from ResolveResult before layout
Summary:
The measure result cache can be used by multiple layout simultaneously because the same ResolveResult can be used to compute multiple layouts. This is a problem Litho's LayoutResult is not reusable (it can also be mutated during collect results today).

Therefore, the MeasureResultCache should only be used by the first layout that picks up the ResolveResult. In order to do this the cache is copied into a new cache, and original one is purged when a new layout starts.

Reviewed By: astreet

Differential Revision: D41769102

fbshipit-source-id: 4681640b4ccf5a68a4a98cfa48287deae76cd7b6