Change viewpager helper section to scroll to initial page index immediately when data bound
Summary:
There is a small time frame after the view pager is visible where the first page is being displayed despite the initial index being set.

Investigated this a bit and found that this is due to the scrolling to initial index being posted to the next UI update.

Change to scroll immediately so that there won't be this delay

Reviewed By: astreet

Differential Revision: D41721927

fbshipit-source-id: 40291642b79d4f27e70e79f177e876bb0b4dca78