Gets the padding from the background drawable during resolve
Summary:
Gets the padding from the background drawable during resolve to avoid race conditions when getting the padding from the drawable.

Since the same Drawable could be currently in layout and be mounted (from a previous layout pass) there are race conditions in resolving the drawable futures.

This is only a problem when render is split into resolve and layout, and litho reuses the previous ResolveResult to calculate a new layout. In the normal case the Drawable would always be resolved again.

By getting the padding from the Drawable during resolve, it will be skipped during layout, and cannot not race with mount. It will also not race with with the resolve phase because the first resolve cannot race with mount (since there isn't anything to mount yet), and once the drawable futures are resolved it is cached in the drawable therefore avoiding any other races.

Reviewed By: astreet

Differential Revision: D41800965

fbshipit-source-id: b10cb9c9a150a1fb48e91bea0fa8cea946f210a2