Change how we handle RTL/LTR text alignment.
Summary: In this post https://fb.workplace.com/groups/bloks.support/permalink/4088138024563168/ we changed the text direction behaviour but now after this https://fb.workplace.com/groups/bloks.support/permalink/5858689710841315/ it seems that the original idea was to change the text alignment only and not the direction. This diff keeps the text direction but will respect the device local alignment.

Reviewed By: pasqualeanatriello

Differential Revision: D41776245

fbshipit-source-id: 49a6e4fd5648295d8142deae48a986ae7272a079