Add abstract test class to force some tests to run in EndToEnd mode
Summary:
I have noticed that the components were not found inside Sections/LazyLists in the `TreeNodeListResolver` when the gradle tests were run in release mode.

I was able to trace it down to this piece of code: https://fburl.com/code/bm3x30nn.

```
    if (!c.isReconciliationEnabled()
        && !ComponentsConfiguration.isDebugModeEnabled
        && !ComponentsConfiguration.isEndToEndTestRun
        && !ComponentsConfiguration.keepLithoNodes) {
      layoutState.mRoot = null;
      layoutState.mLayoutResult = null;
    }
```
It seems that in release mode or when is it is not an end to end run it clears the `layoutState.mLayoutResult`. This way, we are not able to traverse through layout results on sections.

Therefore, I'm adding a temporary work-around (in terms of approach, not content) to have a super class that forces the test to run as "end2end". I want to discuss later with the rest of the team about what would be the best approach to achieve the same effect.

However, this is important to land as it enables the continuous development of this testing API.

It is also not dangerous as this code only lives inside the tests for the new testing API.

Differential Revision: D42146274

fbshipit-source-id: 40ea4e130bdb095e28a6c5236903e9cb1b757606