Remove not needed wrapper runnable to trigger scroll request
Summary:
Before we had to make sure we launched the scroll request inside the main thread. For that we used a `Runnable` that was posted to the Main Thread.

Since we know crash when it is requested outside of the main thread we can remove the runnable and call the code directly.

Differential Revision: D41263983

fbshipit-source-id: adbe5d6bf87de1c3bb9d01991639ceb48e61e3ce