Override toString in Canvas models to reduce binary size
Summary: This diff overrides `toString` in mountable canvas models and returns an empty string in order to reduce the binary size of the models. We can't use `DataClassGenerate` annotation because mountable canvas is used in Litho which is an open source project and that annotation is not open source yet. We may replace these empty `toString` implementations with a better approach at some point.

Reviewed By: colriot

Differential Revision:
D42207763

Privacy Context Container: L1152429

fbshipit-source-id: 72244b87c9943afbbeaa2242719a0ceff2b0f947