Website - Features section image alt text displays as [object Object].
I noticed that when I tried to render an image in the project, the alt text was displayed as [object Object] instead of the expected value - title.

I traced the issue back to the code that was setting the alt attribute of the img element. It was passing React Fragment(object) to the alt attribute, rather than a string.

To fix this, I have added another key 'alt' in the features array, with String as it's value and passed in the Props. This resulted in the expected alt text being displayed.
I have also tried rendering alt attribute this way - alt={title.props.children} which worked. But felt that adding a new key-value pair would be more concise and better.