Website - Features section image alt text displays as [object Object]. (#923)
Summary:
**Title:** Website - Features section image alt text displays as [object Object]

**Description:**

![Screenshot_20230102_071959](https://user-images.githubusercontent.com/91380384/210190327-c693c6fd-526e-48e7-8d54-881ec040478e.png)

I noticed that when I tried to render an image in the project, the alt text was displayed as [object Object] instead of the expected value - title.

I traced the issue back to the code that was setting the alt attribute of the img element. It was passing React Fragment(object) to the alt attribute, rather than a string.

To fix this, I have added another key 'alt' in the features array, with String as it's value and passed in the Props. This resulted in the expected alt text being displayed.
I have also tried rendering alt attribute this way - alt={title.props.children} which worked. But felt that adding a new key-value pair would be more concise and better.

Please review and consider merging this fix.

**Steps to reproduce:**

Run the project in a browser
Observe the image alt text being displayed as [object Object]

**Expected result:**

The image alt text should be displayed correctly.

**Actual result:**

The image alt text is displayed as [object Object].

Pull Request resolved: https://github.com/facebook/litho/pull/923

Reviewed By: adityasharat

Differential Revision: D42311648

Pulled By: astreet

fbshipit-source-id: 534a8ad325f22af200bcda2aeae1696ae88d5492