Add LithoViewComponentsTraverser
Summary:
This is an abstraction that allows to traverse the tree of components and perform an lambda on each component and its parent.

This will be used in the next diff so that we can build an effective tree of `TestNode`. This is the basis so that later we can add filters on the hierarchy of the components (hasParent/hasChild/hasDescendant/hasAncestor).

Reviewed By: kingsleyadio

Differential Revision: D42096907

fbshipit-source-id: a106a84d5f7b3f82238e9e900f58f4c45db33fb5