Fix assigning transition owner key to the Component
Summary:
The `Style.transitionKey` method accepts `ComponentContext` as its first argument. This context should be used to assign its `globalKey` as a transition owner on the component.

Prior to this change, `ComponentContext` passed to `Style.transitionKey` was ignored, and the current component's context which is passed as an argument to `applyToComponent` method was used.

The effect was that some of the transitions were not working with Mountable Components or KComponents. Old component built with `Spec` API were working fine because they were using proper context to assing the transition owner key.

This change fixes the issue and adds a unit test for this scenario.

Reviewed By: apowolny

Differential Revision: D42523188

fbshipit-source-id: 0d78a9908d3dff230170406d8ee3ed1b60d390c0