Set up experiments for different layout calculation threading strategies
Summary:
Set up for a couple of experiments to revisit deprecation of the ThreadPoolLayoutHandler.

Two strategies for initial QE (will be 1 QE with 2 test groups):
1) Always use default layout thread - we've run that experiment in mid-2021 and seen some regressions, but I'm keen to revisit - if it's ok that would be preferred way to deprecate and nuke threadpools.
2) Always use single threaded pools - following improvements in News tab [https://fb.workplace.com/groups/1037863406716494/permalink/1128345987668235/], and since other surfaces (most notably Feed) use 1 as parameter for pools (each surface), let's see the app-wide ramifications of such a change.
If this variant performed well (and I speculate it will perform better than always using default thread), it'd be pretty evident some surfaces do run better when they get an additional thread, but having threadpools of size one is not a great way for that implementation and complexity wise... If this plays out well, short term I'd ship it, and medium term (e.g. after reaching out to stakeholders like Google Elements), I'd strive to deprecate and eventually nuke threadpools, and replace them with an API like ~useOwnLayoutThread() - we could be more flexible with that approach, like making it a HandlerThread and have proper cleanup strategies (I remember memory-wise there were problems with how threadpools are cleaned up).

Reviewed By: astreet

Differential Revision: D42537570

fbshipit-source-id: 4ecf295d3dcf59d8a57acd1c5b8d795bf41d8f30