Fix the crash during dispatching RenderCompleteEvent inside RecyclerBinder
Summary:
**Problem**
Since we introduced `RecyclerBinderAdapterDelegate` for IG adoption, which allows us to be able to customize the root view of each item in RecyclerBinder. However, we didn't change the way notifying render complete, which end up with an exception because `RecyclerView` couldn't find corresponding position through LithoView(it's not the root view).

https://www.internalfb.com/code/fbsource/[a70b17304621582845029e8e28c04464a55aadc5]/fbandroid/libraries/components/litho-widget/src/main/java/com/facebook/litho/widget/RecyclerBinder.java?lines=3932-3942

**Mitigation**
Use `holder.itemView` instead.

Reviewed By: fabiocarballo

Differential Revision: D42554098

fbshipit-source-id: 0977517129dc1d510d77ab7ff96471bd10e81ba2