Gets the LayoutOutput using Extension Input
Summary:
Gets the LayoutOutput using Extension Input and NOT from the view attribute binder.

The ViewAttributeBinder is being added to the Mountable during the layout phase, this is a problem because a Mountable can be reused from the resolve phase which means there can be race between reading from the Mountable being mounted on the main thread and add the binder being added in the background thread. This can cause concurrent modification exceptions.

The most ideal solution would be to add the view attribute binder in the resolve phase, but that is not really possible right now. The view attribute extension sets some properties (padding and layout direction) which are computed in the layout phase. So it is no possible to move this step to the resolve phase.

The workaround is to use extension state instead. A map is created to host the LayoutOutputs, and this is used as the input that is passed to the view attribute extension before mount. Then the extension looks up the LayoutOutput from the map instead (for context this is how incremental mount extension works).

Reviewed By: apowolny

Differential Revision: D42603908

fbshipit-source-id: fa05a5f41b8ef0ef16c94d36ab127401bb3885d0