Do not cull RenderTreeNodes that are zero sized.
Summary:
A RenderTreeNode can be zero-sized and still have children that are non-zero size.  This happens e.g. if a Flexbox does not have a specified size and all its children are absolutely positioned.  Culling the node for this flexbox will mean we end up not rendering the abs positioned children.

See https://fburl.com/pj3xrxk4 for bug report

Reviewed By: pasqualeanatriello

Differential Revision: D42612515

fbshipit-source-id: 60ca96d6a9584adfec404a99c09066b17cbba8c5