Add ability to pass on LifecycleOwner to CT in ComponentTreeHolder
Summary:
While integrating `useLiveData` in D42503749, we have discovered a crash saying the `LifecycleOwner` was not set.

What was happening is while using the `RecyclerBinder`,  the `ComponentTreeHolder` will set-up a `LifecycleProvider` that wraps the original one from the `RecyclerBinder`. However, the information about the `LifecycleOwner` does not propagate through. Hence, when the `Component` tries to access the `LifecycleOwner` it won't find it.

To fix this I created the abstraction of the `AOSPLifecycleOwnerProvider`. This is to be used by any `LifecycleProvider` that has access to a `LifecycleOwner`. At this point we will have two `LithoLifecycleProvider` implementing this:

1. `AOSPLithoLifecycleProvider`: This lifecycle provider is built on top of a `LifecycleOwner` and therefore can provide it to the `ComponentTree` and associate it to it.
2. `ComponentTreeLifecycleProvider`: This lifecycle provider wraps the one of the parent `ComponentTree` that creates it. It will return the `LifecycleOwner` associated to the parent Lifecycle Provider, if this has one.

{F847391875}

Reviewed By: pentiumao

Differential Revision: D42531549

fbshipit-source-id: 3fd79702bdba2931c873811d2f76098892763d1a