Improve error messages
Summary:
Improving error messages shown during selection and matching.

The goal is to have rich messages that will not only notify the error, but also help a developer to carefully debug why the error happened by providing rich contextual information about the state of the selection around the error.

There's a slight change in the behavior of TestNodeSelection#selectChildAt: we don't fail eagerly anymore when the index is out of bounds. Instead, we fail at the point of materialization. This makes it possible to execute a `assertDoesNotExist` on the selection without worrying about failing too early

Reviewed By: fabiocarballo

Differential Revision: D42607727

fbshipit-source-id: 5dbe53510c7468c07127bb45c4ac9f92a61adb71