Make it possible to add fixed binders to RenderUnit
Summary:
Dynamic Binders are Binders that can come and go, and we need to use an identity (`getClass()`) to correlate them between updates. For example, it's possible that a Dynamic Binder was there before and isn't now (or vice-versa).

Fixed Binders, on the other hand, is a concept of a fixed list of Binders that are always there, and they're always in the same order. Fixed Binders don't require an identity, which means we don't have to create a bunch of maps and lists on the mounting/updating hotpath.

Most of the real life usages corresponds to Fixed Binders, an example of Binders that need to be dynamic are Bloks extensions.

Prior to this change, only Dynamic Binders were supported. This change introduces a way to have Fixed Binders in `RenderUnit`.

Reviewed By: astreet, colriot

Differential Revision: D42673857

fbshipit-source-id: d5a96340af74b6f3f1132e78ff179dd9d6b786f7