Expose a missing doc of Sections API
Summary:
This is an on-call [request](https://fb.workplace.com/groups/litho.support/permalink/3451752081812802/), which reminds us the Javadoc of `OnViewportChanged` is outdated and this doc (**communicating-with-the-ui**) is still helpful for users who is using Sections API, just move it to the directory of sections to increase visibility.
A follow-up diff will update the Javadoc of `OnViewportChanged`.

Differential Revision: D42923258

fbshipit-source-id: 06a31bd9196e8f7ed239b5965a9088261c0cb80c