Remove exhaustive extension
Summary:
The exhaustive extension was added to trick the Kotlin compiler to think that `when` is used as an expression. In this case Kotlin compiler enforced such `when` to be exhaustive (to contain all possible values or an `else` clause).

Since Kotlin 1.7, `when` is always exhaustive: https://youtrack.jetbrains.com/issue/KT-47709

Meta is already using Kotlin 1.7.21 https://fb.workplace.com/groups/kotlinandroidfyi/posts/866872077910684/ so it's safe to remove the `exhaustive` extension from Litho codebase.

Reviewed By: fabiocarballo

Differential Revision: D43079252

fbshipit-source-id: 34746a9559968f338eabd80af76d8a9a6bc70f7e