Preload Yoga config on background thread at startup
Summary:
The idea is to provoke Yoga's YogaSetup java class initialization to run early at startup and on background thread, so SoLoader starts loading yoga_internal (which happens in YogaSetup static initialization block, so will run as soon as YogaSetup is first loaded by the JVM, aka the class is first "touched").

We have an ANR which could potentially benefit from this. Putting this behind MC to assess app-wide impact.

Reviewed By: adityasharat

Differential Revision: D43126648

fbshipit-source-id: 09f6fc9d15804c07bc0d2745f6d8f71df464d965