Passes in inter stage props to range API
Summary:
This diff fixes a bug where the range lifecycle methods do not set inter stage props because the interstage props container is never passed to the API. So the AP was setting the container to null. This would throw an NPE when invoking a range lifecycle with inter stage props.

This fix
* The interstage props are held in the RangeTuple.
* Passed to the callbacks as method args.
* The AP and Component APIs have been updated to support this use case.

Differential Revision: D43246229

fbshipit-source-id: a840c1608a3fd06b64e360467bc811cbe2a8d22a