Map common props to attributes
Summary:
So that we can have a generic approach to render all attributes in the `debug dump` - we need to get access to all the test node data in a generic way. This will be done by accessing attributes.

The first step is to then expose the common props data we are using as attributes in the `TestNode`.

Differential Revision: D43273583

fbshipit-source-id: d50639a6fd83d869af9c427deb9e52f352e6b879