Add support to Action specific AttributeKeys
Summary: In order to identify action related attributes such as click, long click, etc in the debug dump, and print them all together, we need to be able to differentiate them from others. In this diff I'm creating an `ActionAttributeKey` which allows us to process these attributes differently in the `Debug`.

Differential Revision: D43273586

fbshipit-source-id: 6467659289984cbd330c248da416bd044cfb5a35