Fix the thread safe issue in CustomSpeedLinearSnapHelper
Summary:
There is a warning reported by infer while removing deprecated constructor of `GridRecyclerConfiguration`, quote here:

> There may be a Thread Safety Violation: Unprotected write. Non-private method CameraRollComponentSpec.onCreateInitialState(...) indirectly writes to field widget.CustomSpeedLinearSnapHelper.mDeltaJumpThreshold outside of synchronization.
Litho components are required to be thread safe because of multi-threaded layout. Reporting because current class is annotated LayoutSpec, so we assume that this method can run in parallel with other non-private methods in the class (including itself).

`mDeltaJumpThreshold` should not be static, change it to a final field.

Reviewed By: fabiocarballo

Differential Revision: D43346735

fbshipit-source-id: b8d96c91b321072c69c9c46c8c6d71a540ba6e75