Make Text an attribute represented as a list of CharSequence
Summary:
In the future, we will need to be able to inspect mounted content and create attributes according to rendered views. For example, if a `Mountable` is inflating a custom view that has two TextViews, we should be able to provide the `Text` information in the `TestNode` for that `Mountable`.

This way, we need to prepare the `Attribute.Text` to become a `List<CharSequence>` instead of a single one.

Differential Revision: D43273580

fbshipit-source-id: 478a563eb826cad40c87d4f5f657cbe2661ef124