Collects the attachables after resolve and commits it with ResolveResult
Summary:
To fix the broken usage of attachables with render to null component instead of walking the LayoutResult tree (which will skip NullNodes) the node tree is traversed after resolve, and the attachables are collected and committed with ResolveResult.

These are then passed to the layout phase. To handle attachables created during layout the nest node trees are traversed after they are laid out.

Reviewed By: astreet

Differential Revision: D42885459

fbshipit-source-id: 23506e8be10fe914d50f2cf1a71b2de507c4645b