Make deps required in useEffect
Summary:
Currently the `useEffect` hook can be called without providing `deps`. In this case `useEffect`'s `onAttach`/`onDetach` is executed on each layout.

This diff changes this behavior by not allowing to call `useEffect` without `deps`.

This way the API is more restrictive and the developers have to be explicit when they want to call `useEffect` and they don't want to provide any meaningful `deps`.
This also makes this API similar to [Jetpack Compose side effects](https://developer.android.com/jetpack/compose/side-effects). It should be easier now for folks who know Compose to understand `useEffect`.

I'll be updating the `useEffect` documentation once this change lands. But the main patterns will be:
1. passing correct `deps` to `useEffect` - in this case `useEffect`'s `onAttach`/`onDetach` is executed whenever any of the `deps` changes between layouts.
2. passing `Unit` as a `dep` like `useEffect(Unit) { ... }` - in this case `useEffect`'s `onAttach` is executed once on initial mount and `onDetach` is executed once on final unmount.
3. passing `Any()` as a `dep` like `useEffect(Any()) { ... }` - in this case `useEffect`'s `onAttach`/`onDetach` is executed on each layout. This behavior is the same as calling `useEffect` without `deps` prior to the changes introduced in this diff.

This diff also removes `usePersistentEffect` which is not necessary anymore, because the same behavior can be achieved with `useEffect(Unit) { ... }`. And it also updates all callsites that called `useEffect` without `deps` or used `usePersistentEffect`.

Reviewed By: astreet

Differential Revision: D43431037

fbshipit-source-id: f80a74e5ab1bcf38dbfd628b1bb9f8d5cc16ad00