Re-enable litho-testing with Gradle
Summary: The tests that were faulty were only happening with Gradle. This seems to indicate only an inconsistency in the LooperMode - which was the factor that changed before. As it was fixed in D43463782 (https://github.com/facebook/litho/commit/d32014192beb864d53d5cd188bbd2a2134a90319), I believe we can re-enable them.

Reviewed By: adityasharat

Differential Revision: D43463783

fbshipit-source-id: 2d3919f929d06f1ff18ad8a34a699277ef1fe85b