Run all tests against latest SDK (33)
Summary:
This diff makes all litho tests to be ran against the latest SDK version we compile into, which is 33.

Due to those changes some tests broke and I had to force some tests to still run on older SDKs. We should revisit them as soon as we can so that they can also be ran against 33.

Reviewed By: adityasharat

Differential Revision: D43457251

fbshipit-source-id: 5047ee4fc33d9148254684fb25177b73dec27511