Remove not used robolectric jars
Summary:
As we were aligning Robolectric versions between Gradle and BUCK there was a need to add new jars under `lib/fb/android-all` (to make the `litho-oss-test` task pass).

This diff focused in cleaning up all the old jars that are no longer needed.

The total wins (between the jars added and removed) are of around 195

Reviewed By: pentiumao

Differential Revision: D43499594

fbshipit-source-id: b799dc29b1e2781bc46a50631255af73778c073b