Allow RenderResultFuture to take in a Callable
Summary:
For the new collection measurement algorithm, a single item can be measured twice in a single layout pass to satisfy measurement constraints like minimum cell sizes or aspect ratios or stretching along the cross axis.  We want this optional double measurement to happen on the background thread.

Right now, a RenderResultFuture represents a single call to render() that can be scheduled to the background.

Exposing a constructor that can take in a Callable allows us to have RenderResultFuture that would be able to encapsulate the above optional double measurement.

Reviewed By: zielinskimz

Differential Revision:
D43313894

Privacy Context Container: L1152429

fbshipit-source-id: a87d726c0e9424a41a17990eac544f5dbfc29b5d