Add text input and update character count
Summary:
This diff uses the prebuilt Kcomponent `TextInput` to implement the text input field. The KComponent does not provide all the properties as the Java spec API does. I added the `inputBackground` for the KComponent version, or it will add some paddings to the input text component, making the text not vertically centered in our bubble.

To update the character, I use the Hooks `useState` and pass the event handler to the child component.

Differential Revision: D43643585

fbshipit-source-id: 5629bc82c2a07ea624aec84fcd58cf2a39d0f819