Integrate resolve cancellation in ComponentTree
Summary:
This is a series of diffs that are focused in providing support for a first iteration of *Resolve* cancellation. In the schema below you can find a high-level overview of the changes required and the end flow.

This flow will also be possible to be used when we move for specific *Layout* cancellation.

{F886248253}

 ---

In this diff we focus on integrating the `ResolveCancellationEvaluator` with `ComponentTree#doResolve` step.

Reviewed By: adityasharat

Differential Revision: D43618753

fbshipit-source-id: eb0efaa87bcec9b35a67c7444a21d3b90ccc2489