Add infra-structure for CancellationEvaluator and concrete implementation for Resolve
Summary:
This is a series of diffs that are focused in providing support for a first iteration of *Resolve* cancellation. In the schema below you can find a high-level overview of the changes required and the end flow.

This flow will also be possible to be used when we move for specific *Layout* cancellation.

{F886248253}

In this specific diff I implement the `CancellationEvaluator` abstraction and the concrete implementation for *Resolve*.

Whenever we would want to support Layout cancellation we would would have to do a similiar diff to this one (and corresponding integration) - but the core mechanism could be the same.

The logic for *Resolve* cancellation follows these lines:

{F886251887}

Reviewed By: adityasharat

Differential Revision: D43618754

fbshipit-source-id: 8bda6c5636795dbb950b58ab7b5de7c5004fb70a