Gate Resolve Cancellation with ComponentsConfiguration
Summary:
This is a series of diffs that are focused in providing support for a first iteration of *Resolve* cancellation. In the schema below you can find a high-level overview of the changes required and the end flow.

This flow will also be possible to be used when we move for specific *Layout* cancellation.

{F886248253}

----

In this diff we make the *Resolve Cancellation* optional according to the `ComponentsConfiguration` setting.

Reviewed By: adityasharat

Differential Revision: D43618752

fbshipit-source-id: 4892482c19084855f8b62309d279aa553cb10af4