Simplify PrimitiveLayoutResult
Summary:
This diff simplifies `PrimitiveLayoutResult` public API. Previously it was required to pass widthSpec, heightSpec and renderUnit when creating PrimitiveLayoutResult. This is not needed anymore, the Primitive API passes these values internally so that the developer doesn't have to do it.

Also prior to this change, the renderUnit was never properly assigned to PrimitiveLayoutResult returned from LayoutBehavior.layout() function. This has been fixed as well.

Reviewed By: adityasharat

Differential Revision: D43864676

fbshipit-source-id: c761a0463b965d97a225bf3333a9b1afbc2fa758