Add viewId Style method
Summary:
In this diff we add a new `Style` method that allows setting a `viewId` to the associated `Component`. If the `Component` will be represented as a `View` (this is, wrapped in one) if it wasn't already at `mount` time.

This will help to support a use case for IG4A, where some frameworks still rely in the ids of views. (cc dmitriykarmazin)

Reviewed By: zielinskimz

Differential Revision: D44057483

fbshipit-source-id: fa37815d88a61b21ef44313a8c283ac2dfd51ed8