Add ability to associated a breadcrumb with Litho debug logs
Summary:
In this diff we introduce a log "breadcrumb" for the `ComponentTree` debug logs. If this breadcrumb is set, it will include it in the log messages.

This is extremely helpful when you are attempting to filter out the logs for a given ComponentTree (associated with a specific identifier). I have used this approach in the last days while debugging IG4A issues and it helped me quite a lot.

Reviewed By: zielinskimz

Differential Revision: D44097170

fbshipit-source-id: f1ddab3825456d093c386d572c40d32c4243eba3