Remove dependency on TreeProps from ComponentsLogger
Summary:
This is the 3rd diff of this stack that has the goal to enable QPL logging using ComponentsLogger widely on FB4A.

At the moment, this ComponentLogger abstraction lives inside litho-core and has explicit dependencies on litho internals (such as TreeProps and ComponentContext). In order to enable a global configuration (via ComponentsConfiguration) we need to extract these abstractions to another module into which litho-core and litho-configurations can depend on.

These are the steps:
- Create module for performance logger code ✅
- **Remove litho internals dependencies from performance logger code.** ✅ (focus of this diff)
- Move litho performance logger code into the created module. 🔜
- Create global configuration in ComponentsConfiguration that is used whenever there is no local logger provided. 🔜

Reviewed By: pentiumao

Differential Revision: D44105597

fbshipit-source-id: f16b50be8a556f654ea1d2b450257a54201e9b9e