Move ComponentsLogger and PerfEvent to litho-perf-logger module
Summary:
This is the 4th diff of this stack that has the goal to enable QPL logging using ComponentsLogger widely on FB4A.

At the moment, this ComponentLogger abstraction lives inside litho-core and has explicit dependencies on litho internals (such as TreeProps and ComponentContext). In order to enable a global configuration (via ComponentsConfiguration) we need to extract these abstractions to another module into which litho-core and litho-configurations can depend on.

These are the steps:
- Create module for performance logger code ✅
- Remove litho internals dependencies from performance logger code. ✅
- **Move litho performance logger code into the created module.** ✅ (focus of this diff)
- Create global configuration in ComponentsConfiguration that is used whenever there is no local logger provided. 🔜

Reviewed By: pentiumao

Differential Revision: D44105598

fbshipit-source-id: 2a11ca9d606f5de94ad8a7c13becfc023b3f7273