Add ComponentsLogger to ComponentsConfiguration
Summary:
This is the 5th diff of this stack that has the goal to enable QPL logging using ComponentsLogger widely on FB4A.

At the moment, this ComponentLogger abstraction lives inside litho-core and has explicit dependencies on litho internals (such as TreeProps and ComponentContext). In order to enable a global configuration (via ComponentsConfiguration) we need to extract these abstractions to another module into which litho-core and litho-configurations can depend on.

These are the steps:
- Create module for performance logger code ✅
- Remove litho internals dependencies from performance logger code. ✅
- Move litho performance logger code into the created module. ✅ (focus of this diff)
- Create global configuration in ComponentsConfiguration and integrate global ComponentsLogger with ComponentContext ✅ (focus of this diff)

Reviewed By: zielinskimz

Differential Revision: D44168460

fbshipit-source-id: 6eaebbcd42164ab777458cef73299eab58c400ec