Back out "Back out "commit resolve and layout separately""
Summary:
This is a fixed version of the stack that was reverted in D43274289.
I identified the problem with the original diff and verified that none of the issues that were being signaled are now reproducing.
I will add a test before landing the diff

Original commit changeset: 8001c11d1f7d

Original Phabricator Diff: D43274289

Reviewed By: colriot

Differential Revision: D44292342

fbshipit-source-id: 2053781fa4f2cf9eadf3c15d6fdc3ba4a981edec