Unset view id during unmount
Summary:
This diff unsets the view id during unmount.

The previous attempt (D44365959) was backed out because the ID was unset during unmount even if it hasn't been set via Style.viewId API. It means that for example, when someone had a component and in `createContent()` they returned a view that already has an id, then D44365959 would unset that id during unmount.
D44365959 broke a lot of tests, so this time I made sure that I check the `Attributes` before unsetting the id. If the id flag is set inside of `Attributes` then it means that the id has been set via Style.viewId API and it's safe to unset it during unmount.

Reviewed By: fabiocarballo

Differential Revision: D44409444

fbshipit-source-id: 212f30bc8024a823971fa30922a9d4bc611ab941