Add SizeSpecsMismatchDetectorComponent to help detect integration issues in Litho migration
Summary:
In this diff we introduce a `SizeSpecsMismatchDetectorComponent` as a wrapper of the Feed Litho Items. This `Component` will detect (only in debug mode) inconsistencies in the size specs declared and requested by `LithoView#measure` calls and if so, it will update the background color of the given `Component` where the mismatch happens.

The color of the rendering can be controlled by overriding the `debugMismatchColor` inside a specific `LithoBinder` implementation. If nothing is specified, it will default to `red`.

Reviewed By: zielinskimz

Differential Revision: D44420261

fbshipit-source-id: 43fba74688d0ef16392dcb62bd690457a8271062