Remove unused metadata API
Summary:
The metadata API in `Component.java` was introduced in: D29850351

However, at some point later, in this diff: D31347774 the calls to `getMetadata` method were removed. Since `getMetadata` was not used anymore, it has been removed in: D33816647

Right now it looks like there is a way to add metadata but not to get it back.

The only places that still add metadata are:
 - `PermalinkTopLevelFooterSelectorComponentSpec.java`
 - `PermalinkUFIFeedbackSummarySelectorComponentSpec.java`

This diff removes the remaining metadata API parts from `Component.java` and removes the calls from the two files mentioned above.

Reviewed By: adityasharat

Differential Revision: D44464714

fbshipit-source-id: c9a1d93b678a4b3d0c3a317e847490be9ceb9383