Add ability to detect main thread renders in Litho x Feed components
Summary:
In this diff I had to re-think a bit the initial concept for tooling. The goal was to get to provide a visual hint whenever a main thread render happened for a Component. In the context of our integration it can be usually for one of two reasons : (1) size specs mismatches and (2) no main thread layout state commited yet.

I ended up on creating a `LithoMainThreadRenderSourcesDetector` , which listens to debug events from Litho and passes on that information to the `LithoBinder`. The `LithoBinder` then gets hold of the correct preparable and renders an overlay on the associated `LithoView`.

I have put this behavior behind a developer option. However, it does require a restart of the app once the setting is selected.

Reviewed By: adityasharat

Differential Revision: D44464205

fbshipit-source-id: bf83b7f759b660650ee432fb2c14382b96e7c1d2