Add flag to control if LithoView unmounts all items when Preparable is cleared
Summary:
In my understanding there is one further optimization that we can do in this scenario. Every time we `release` the `ComponentTree` when the `Preparable` is cleared, it goes through a flow where it will unmount all items from the `LithoView`. Also, in theory if something is getting out of the window, then its items should have been already been unmounted, so there is no reason that we go through that flow that seems to be heavy.

I have did some local debugging and by avoiding the unmount I the `clear` went from almost 0.9ms to 0.1 in average. I haven't noticed any noticeable different behaviors, so I'm putting this behind a flag that we can experiment in Beta.

Reviewed By: adityasharat

Differential Revision: D44720147

fbshipit-source-id: d58115924d3bd0e0358e9198b3ca0c367eeaaf2a