Add ability to share renderUnitIdGenerator
Summary:
This renderUnitIdGenerator is meant to be shared by all `ComponentTree` created for the binder for the specific `itemType`.

By sharing the `RenderUnitIdGenerator` we are able identify the same `RenderUnit` across different `ComponentTree`. This allows for a more efficient mounting process, as the same Component Hierarchy will generate the same `RenderUnit` ids. For example, two different `MediaUfiComponent`, with the same Component Hierarchy, will create `RenderUnit` with the same ids.

These same ids will be used during the mount process. Ideally, even re-using a [LithoView], the new `ComponentTree` will attempt to mount the same `RenderUnit` in a lot of occasions - avoiding a full mount/unmount cycle, and simply updating when needed.

When this is used, we should also create the `LithoPreparable` with the `LithoPreparable.unmountOnClear` set to `false`, to guarantee that `MountItem` are not unmounted, and that we can simply re-use them when we bind another `ComponentTree` with the same type of hierarchy.

Reviewed By: adityasharat

Differential Revision: D44745487

fbshipit-source-id: 44f7189b82530deee357932c9b169b02b31d531e